+++ forten/src/pages/onboard/cubepage/index.tsx
