+++ forten/src/pages/onboard/cubepage/styles.css
@@ -1,19 +1,17 @@
 
 .mod {
-    position: fixed;
-    top: 0px;
+    /* position: fixed; */
+    bottom: 0px;
     left: 0px;
     width: 100%;
-    height: 100%;
+    height:100vh;
     perspective: 330px;
-    overflow: hidden;
-    background: radial-gradient(
-    220% 105% at top center,
-    rgb(40, 61, 107) 10%,
-    #0d0d23
-  );
-   
+    overflow-y: hidden;
+    background: linear-gradient(180deg,#4e6085 0%, #171C26 50%);
+  
+
   }
+
   .mod .project-name {
     position: absolute;
     width: 420px;
