+++ forten/src/pages/onboard/wavepage/styles.css
@@ -21,11 +21,12 @@ p {
   position:relative;
   text-align:center;
   /* background: linear-gradient(60deg, rgb(69, 69, 70) 0%, rgb(149, 149, 149) 100%); */
-  background: radial-gradient(
+  /* background: radial-gradient(
     220% 105% at top center,
     #1b2947 10%,
     #000000
-  );
+  ); */
+  background: radial-gradient(220% 105% at top center, #4e6085 10%, #ffffff);
   background-attachment: fixed;
   overflow: hidden;
 
@@ -35,7 +36,7 @@ p {
 
 
 .inner-header {
-  height:65vh;
+  height:30vh;
   width:100%;
   margin: 0;
   padding: 0;
@@ -51,10 +52,10 @@ p {
 .waves {
   position:relative;
   width: 100%;
-  height:15vh;
-  margin-bottom:-7px; /*Fix for safari gap*/
-  min-height:100px;
-  max-height:150px;
+  height:60vh;
+  margin-bottom:-70px; /*Fix for safari gap*/
+  /* min-height:100px; */
+  max-height:50vh;
 }
 
 .content {
